Fix crashes from template finalizer (#182)
* Fix crash from template finalizer releasing V8 data, just free the wrapper

v8::Persistent::Reset() shouldn't be used without synchronization
from the finalizer goroutine to avoid race conditions. Instead, we can
just free the wrapper and leave the V8 data on the isolate heap to be
freed when the isolate is disposed.

* Keep alive the template while its C++ pointer is still being used

Otherwise, the garbage collector can collect the template and the
finalizer can free the C++ m_template struct before it is finished
being used.